refactor: #12 cleanup codes.
