+++ .gitconfig.sample
@@ -1,3 +1,3 @@
 [user]
-	name = Your Name
-	email = yourmail@example.com
+name = Your Name
+email = yourmail@example.com
