+++ Cargo.toml
@@ -8,10 +8,10 @@ publish = false
 
 [dependencies]
 axum = "0.6.20"
-tokio = { version = "1", features = ["full"] }
-hyper = { version = "0", features = ["full"] }
-tower = "0"
-rand = "0"
+tokio = { version = "1.34.0", features = ["full"] }
+hyper = { version = "0.14.27", features = ["full"] }
+tower = "0.4.13"
+rand = "0.8.5"
 serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-sysinfo = "0.29"
+serde_json = "1.0.108"
+sysinfo = "0.29.10"
