+++ src/controllers/dice.rs
@@ -8,7 +8,7 @@ struct Dice {
     dice: i32,
 }
 
-pub async fn dice() -> impl IntoResponse {
+pub async fn roll() -> impl IntoResponse {
     let mut rnd: rand::rngs::ThreadRng = rand::thread_rng();
     let i: i32 = rnd.gen_range(1..6);
     let dice = Dice { dice: i };
