+++ src/controllers/metrics.rs
