+++ src/controllers/mod.rs
@@ -1,5 +1,5 @@
 pub mod root;
 pub mod dice;
-pub mod make_sleep;
-pub mod get_metrics;
+pub mod sleepy;
+pub mod metrics;
 pub mod bar;
